+++ src/module.ts
@@ -1,13 +1,28 @@
+import type { Config } from 'drizzle-kit'
 import { existsSync } from 'node:fs'
-import { addServerImports, addServerPlugin, addServerTemplate, createResolver, defineNuxtModule, resolvePath, useLogger } from '@nuxt/kit'
+import { addServerImports, addServerPlugin, addServerTemplate, createResolver, defineNuxtModule, findPath, resolvePath, useLogger } from '@nuxt/kit'
 import { globby } from 'globby'
+import { createJiti } from 'jiti'
 import { basename, dirname, join } from 'pathe'
 import { name, version } from '../package.json'
 
+const logger = useLogger('drizzle-migrations')
+
 export interface ModuleOptions {
   /**
-   * Path to the directory containing migration files
-   * @default 'server/database/migrations'
+   * Path to the `drizzle.config.ts` file used for auto-detecting migration settings.
+   *
+   * Will auto-detect one of `['drizzle.config.ts', 'drizzle.config.js', 'drizzle.config.json']`
+   *
+   * @default auto-detect
+   */
+  configPath: string
+  /**
+   * Path to the directory containing migration files.
+   *
+   * If omitted, will try to read from {@linkcode configPath}.
+   *
+   * @default 'drizzle'
    */
   migrationsPath: string
   /**
@@ -17,7 +32,7 @@ export interface ModuleOptions {
   storageName: string
 }
 
-export default defineNuxtModule<ModuleOptions>({
+export default defineNuxtModule<ModuleOptions>().with({
   meta: {
     name,
     version,
@@ -27,13 +42,14 @@ export default defineNuxtModule<ModuleOptions>({
     },
   },
   defaults: {
-    migrationsPath: 'server/database/migrations',
     storageName: 'migrations',
   },
   async setup(options, nuxt) {
-    const logger = useLogger('drizzle-migrations')
     const { resolve } = createResolver(import.meta.url)
-    const migrationsPath = await resolvePath(options.migrationsPath)
+    const configPath = await resolveDrizzleConfigPath(options.configPath)
+    const drizzleConfig = configPath ? await readDrizzleConfig(configPath) : undefined
+
+    const migrationsPath = await resolvePath(options.migrationsPath ?? drizzleConfig?.out ?? 'drizzle')
     const migrationFolderVersion = await getMigrationFilesVersion(migrationsPath)
     if (migrationFolderVersion === null) {
       logger.error('Could not determine migration folder version. Check that the migrationsPath is correct and contains valid migration files.')
@@ -63,11 +79,12 @@ export default defineNuxtModule<ModuleOptions>({
     addServerTemplate({
       filename: '#drizzle-migrations',
       getContents: () => `
-        export { default as journal } from '#drizzle-migrations/journal';
-        export const storageName = ${JSON.stringify(options.storageName)}
+export { default as journal } from '#drizzle-migrations/journal'
 
-        export const migrationFolderVersion = ${JSON.stringify(migrationFolderVersion)};
-        `,
+export const storageName = ${JSON.stringify(options.storageName)}
+
+export const migrationFolderVersion = ${JSON.stringify(migrationFolderVersion)};
+`,
     })
 
     addServerTemplate({
@@ -85,6 +102,28 @@ export default defineNuxtModule<ModuleOptions>({
     addServerPlugin(resolve('./runtime/server/plugin'))
   },
 })
+
+/**
+ * Resolves the Drizzle config path. If no path is provided, it attempts to auto-detect it.
+ */
+async function resolveDrizzleConfigPath(configPath?: string) {
+  return configPath
+    ? await resolvePath(configPath)
+    : await findPath('drizzle.config', {
+        extensions: ['.ts', '.js', '.json'],
+        type: 'file',
+      })
+}
+
+async function readDrizzleConfig(configPath: string) {
+  const jiti = createJiti(import.meta.url)
+  try {
+    return await jiti.import<Config>(configPath, { default: true })
+  }
+  catch (e) {
+    logger.error(`Failed to read Drizzle config at ${configPath}: ${(e as Error).message}`)
+  }
+}
 
 /**
  * Determine the version of migration files in the given path.