+++ src/runtime/server/drizzle-migrations.d.ts
@@ -1,10 +1,17 @@
 declare module '#drizzle-migrations' {
-  export const journal: typeof import('#drizzle-migrations/journal')['default']
+  export { default as journal } from '#drizzle-migrations/journal'
+
   export const storageName: string
   export const migrationFolderVersion: 2 | 3
 }
 
 declare module '#drizzle-migrations/journal' {
-  const journal: import('./readMigration').Journal
+  const journal: {
+    entries: {
+      when: number
+      tag: string
+      breakpoints: boolean
+    }[]
+  }
   export default journal
 }