+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## v0.2.0
+
+[compare changes](https://github.com/mattmess1221/nuxt-drizzle-migrations/compare/v0.1.1...v0.2.0)
+
+### 🚀 Enhancements
+
+- Load default `migrationsPath` from drizzle config file ([f92615a](https://github.com/mattmess1221/nuxt-drizzle-migrations/commit/f92615a))
+
+### 💅 Refactors
+
+- Load the journal file as an import alias ([d5b96b6](https://github.com/mattmess1221/nuxt-drizzle-migrations/commit/d5b96b6))
+
+### 🌊 Types
+
+- Add missing ts-ignore for server imports ([ac2c8cf](https://github.com/mattmess1221/nuxt-drizzle-migrations/commit/ac2c8cf))
+
+### ❤️ Contributors
+
+- Matthew Messinger <mattmess1221@gmail.com>
+
 ## v0.1.1
 
 [compare changes](https://github.com/mattmess1221/nuxt-drizzle-migrations/compare/v0.1.0...v0.1.1)